repo-root/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
fastapi
uvicorn
requests
python-dotenv
pydantic
openai
beautifulsoup4
PyPDF2
python-docx
dropbox
